# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Supply Checkout API
  version: 1.0.0
  summary: Team data (products and sheets), the signed-in user's teams and invites, and sign-in sessions for the web app.
  description: |
    The HTTP API in front of the data Lambdas (ADR 0006). The browser adapter
    (supply-checkout-a2b) implements the app's `window.claude.use("db")` on
    these routes; the handlers are in `backend/src/api/` and the route table
    they share with the CDK app is `backend/src/api/routes.ts`. A test
    (`backend/test/openapi.test.ts`) keeps this file and that table in step.

    ## Mapping the app's db calls

    | App call (src/main.js) | Request |
    | --- | --- |
    | `db.collection("products").onSnapshot(cb)` | `GET /teams/{teamId}/products`, following `cursor` until absent; `snap.docs[i].id` = `id`, `.data()` = `data` |
    | `db.collection("sheets").orderBy("date", "desc").onSnapshot(cb)` | `GET /teams/{teamId}/sheets?orderBy=date&direction=desc`, following `cursor` |
    | `db.doc("<coll>/<id>").get()` | `GET /teams/{teamId}/<coll>/<id>`; 404 `not_found` means `exists: false` |
    | `db.doc("<coll>/<id>").set(data)` | `PUT /teams/{teamId}/<coll>/<id>` with `{"data": data}` |
    | `db.doc("<coll>/<id>").update(patch)` | `PATCH /teams/{teamId}/<coll>/<id>` with `{"data": patch}` (deep merge) |
    | `db.doc("<coll>/<id>").delete()` | `DELETE /teams/{teamId}/<coll>/<id>` |
...

    - **Checkout and return** don't use `PATCH` plus a stock update: they're
...
      team.
...

...

...
servers:
...
security:
  - cognito: []
...
paths:
...
  /teams/{teamId}/products/{key}/stock:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - $ref: "#/components/parameters/productKey"
    post:
      tags: [commands]
      operationId: adjustStock
      summary: Change an item's stock outside a sheet, in one transaction, with a movement in its history.
      description: |
        - `reason: "receipt"`: `quantity` eaches bought at `unitCost` each
          (after any pack conversion, ADR 0014) are added to stock. An item
          that didn't track stock starts at `quantity`. The item's `price`
          and `cost` don't change; update the document for that.
        - `reason: "count"`: someone counted `count` in storage. Stock is set
          to it, and the movement records the difference.

        Contributors and owners only. Idempotent by `operationId`.
      requestBody: { $ref: "#/components/requestBodies/StockAdjust" }
      responses:
        "200": { $ref: "#/components/responses/CommandDone" }
        "400": { $ref: "#/components/responses/CommandRefused" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/CommandConflict" }
  /teams/{teamId}/products/{key}/movements:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - $ref: "#/components/parameters/productKey"
    get:
      tags: [commands]
      operationId: listMovements
      summary: An item's stock history, newest first. Any member can read it.
      parameters:
        - name: limit
          in: query
          description: At most this many movements. Default 50.
          schema: { type: integer, minimum: 1, maximum: 100 }
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: A page of movements.
          content:
            application/json:
              schema:
                type: object
                required: [movements]
                properties:
                  movements:
                    type: array
                    items: { $ref: "#/components/schemas/Movement" }
                  cursor: { type: string, description: Absent on the last page. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /teams/{teamId}/imports:
    parameters: [{ $ref: "#/components/parameters/teamId" }]
    post:
      tags: [products]
      operationId: importProducts
      summary: Import items from a CSV file, all or nothing, with a dry-run preview. Owners only.
      description: |
        The server reads the file (columns `name`, `barcode`, `price`, `cost`,
        `stock`, `pack_size`, matched without regard to case or spacing; other
        columns are left out) and validates every row against the whole
        inventory before it writes anything. A row matches an existing item by
        barcode, or by name when the row or the item has no barcode, and
        updates it; a re-import **sets** price, cost, pack size and stock to
        the file's values and never duplicates an item. Blank optional cells
        keep what's there. Money may have `$` and thousands separators and is
        rounded to cents; stock and pack size are whole numbers (ADR 0014).

        - `dryRun: true`: writes nothing. 200 with `status: "preview"` (or
          `"invalid"` with every problem) and what each row will do.
        - Otherwise: 200 `imported` when every row is in. A file with any bad
          row is 400 `bad_request` with every problem in `errors`, and nothing
          is written. The import is staged, then committed in transactions of
          up to 49 rows and 2.5 MB; if it stops part-way (a timeout, or 409 when
          items kept changing), send the same request again (same `importId`
          and file, from any owner) to finish it. The same `importId` with a
          finished import returns its summary with `replayed: true`; with a
          different file it is 400. A 409 whose message says to choose the
          file again (the import expired, or a key it planned was taken by an
          item with another barcode) needs a new `importId`.

        At most 1,000 rows and 300 KB. Owners only: a contributor or viewer
        gets 403 `permission_denied` with `reason: "owners_only"`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [csv]
              properties:
                importId: { allOf: [{ $ref: "#/components/schemas/OperationId" }], description: "One UUID per import, sent again unchanged on every retry. Required unless `dryRun`." }
                csv: { type: string, maxLength: 300000, description: The file's text (UTF-8). The first non-blank row names the columns. }
                dryRun: { type: boolean, description: "True: validate and preview, and write nothing." }
            examples:
              preview:
                summary: Preview a file
                value: { dryRun: true, csv: "name,barcode,price,cost,stock,pack_size\nNitrile gloves,0123,12.50,9.99,40,100\n" }
      responses:
        "200":
          description: The preview (dry run), or the finished import.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [status, rows, errors, errorCount, ignoredColumns, summary]
                    properties:
                      status: { type: string, enum: [preview, invalid] }
                      rows: { type: array, items: { $ref: "#/components/schemas/ImportRow" }, description: The valid rows and what each will do. }
                      errors: { type: array, maxItems: 200, items: { $ref: "#/components/schemas/ImportRowError" } }
                      errorCount: { type: integer }
                      ignoredColumns: { type: array, items: { type: string } }
                      summary: { $ref: "#/components/schemas/ImportSummary" }
                  - type: object
                    required: [status, importId, replayed, summary]
                    properties:
                      status: { const: imported }
                      importId: { type: string, format: uuid }
                      replayed: { type: boolean }
                      summary: { $ref: "#/components/schemas/ImportSummary" }
        "400":
          description: "`bad_request`, nothing written: a malformed body or file, or rows with problems (listed in `errors`, at most 200, and counted in `errorCount`)."
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Error" }
                  - type: object
                    properties:
                      errors: { type: array, items: { $ref: "#/components/schemas/ImportRowError" } }
                      errorCount: { type: integer }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`aborted`: items kept changing under the import (send the same request again to finish it), or it can't go on as staged: its plan expired, or a key it planned to create was taken by an item with another barcode (import the file again with a new `importId`). The message says which."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "413": { $ref: "#/components/responses/TooLarge" }
...
  /teams/{teamId}/sheets/{sheetId}/lines:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - $ref: "#/components/parameters/sheetId"
    post:
      tags: [commands]
      operationId: addLines
      summary: Add a receipt's lines for a client to a sheet, in one transaction, without moving stock.
      description: |
        Adds each line's `quantity` to the sheet's `out` for its product,
        creating the line, with the request's `code`, `name`, `price` and
        `cost`, if the sheet doesn't have it. An existing line keeps its copy.
        Every line changes or none does. Stock doesn't move: the items were
        bought for the client and never were in storage. 1 to 40 lines, each
        product at most once. The sheet must be open. Contributors and owners
        only. Idempotent by `operationId`: see [commands.md](commands.md).
      requestBody: { $ref: "#/components/requestBodies/AddLines" }
      responses:
        "200": { $ref: "#/components/responses/LinesAdded" }
        "400": { $ref: "#/components/responses/CommandRefused" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/CommandConflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
  /teams/{teamId}/sheets/{sheetId}/return:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - $ref: "#/components/parameters/sheetId"
    post:
      tags: [commands]
      operationId: returnItems
      summary: Record whole eaches of a line coming back, and put them back in stock, in one transaction.
      description: |
        Adds `quantity` to the line's `returned`, which can never go above
        `out`, and adds `quantity` to the product's `stock` if the product
        tracks stock. Prices aren't touched. The sheet must be open: a closed
        sheet takes no returns (reopen it first, or correct the line's counts
        with a `PATCH`, which doesn't move stock). Contributors and owners
        only. Idempotent by `operationId`.
      requestBody: { $ref: "#/components/requestBodies/Return" }
      responses:
        "200": { $ref: "#/components/responses/CommandDone" }
        "400": { $ref: "#/components/responses/CommandRefused" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/CommandConflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
...
components:
  securitySchemes:
    cognito:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Cognito **access token** from the environment's user pool and the web
        app client (`Authorization: Bearer <token>`). API Gateway checks the
        signature, expiry, issuer and client; the Lambda also requires
        `token_use: access` and an unexpired `exp`. ID tokens are refused.
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Secure-sc_refresh
      description: |
        The refresh token, set by `/auth/session` and `/auth/refresh`:
        `HttpOnly; Secure; SameSite=Strict; Path=/auth; Max-Age=2592000`.
        Requests must also carry an allowed `Origin` (CSRF protection).
  parameters:
    teamId:
      name: teamId
      in: path
      required: true
      description: The team. The caller must be a member; the body can't name a team.
      schema: { type: string, pattern: "^[A-Za-z0-9_-]{1,128}$" }
...
    productKey:
      name: key
      in: path
      required: true
      description: The product key, percent-encoded. Up to 256 characters, no `/` or control characters, and not `__proto__`.
      schema: { type: string, minLength: 1, maxLength: 256 }
...
    cursor:
      name: cursor
      in: query
      description: The `cursor` from the previous page, as is. Only valid for the same listing.
      schema: { type: string }
...
  requestBodies:
...
    StockAdjust:
      required: true
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                additionalProperties: false
                required: [operationId, reason, quantity, unitCost]
                properties:
                  operationId: { $ref: "#/components/schemas/OperationId" }
                  reason: { const: receipt }
                  quantity: { $ref: "#/components/schemas/Quantity" }
                  unitCost: { allOf: [{ $ref: "#/components/schemas/Money" }], description: What was paid per each, before tax. }
              - type: object
                additionalProperties: false
                required: [operationId, reason, count]
                properties:
                  operationId: { $ref: "#/components/schemas/OperationId" }
                  reason: { const: count }
                  count: { type: integer, minimum: 0, maximum: 1000000, description: The number counted in storage. }
  schemas:
    OperationId:
      type: string
      format: uuid
      description: |
        A UUID the client makes for one intended action (`crypto.randomUUID()`)
        and sends again, unchanged, on every retry of it. The same ID returns
        the first result for 7 days; the same ID with a different request is
        `bad_request`.
...
    Movement:
      type: object
      description: One change to an item's stock. The sum of an item's `delta`s is its change in stock.
      required: [type, productKey, reason, delta, tracked, operationId, userId, at]
      properties:
        type: { const: movement }
        productKey: { type: string }
        reason: { type: string, enum: [checkout, return, receipt, count, import], description: "`import`: a CSV import set the stock (`count`)." }
        delta: { type: integer, description: The change to `stock`. 0 when the item didn't track stock. }
        tracked: { type: boolean, description: Whether the item tracked stock. }
        quantity: { type: integer }
        count: { type: integer }
        sheetId: { type: string }
        unitCost: { type: number }
        operationId: { type: string, format: uuid }
        userId: { type: string }
        at: { type: string, format: date-time }
    ImportRow:
      type: object
      required: [line, name, barcode, price, key, action, changes]
      properties:
        line: { type: integer, description: The row's line in the file. }
        name: { type: string, maxLength: 200 }
        barcode: { type: string, maxLength: 256 }
        price: { $ref: "#/components/schemas/Money" }
        cost: { $ref: "#/components/schemas/Money" }
        stock: { type: integer, minimum: 0, maximum: 1000000 }
        packSize: { type: integer, minimum: 1, maximum: 10000 }
        key: { type: string, description: The product key the row writes. }
        action: { type: string, enum: [create, update, unchanged] }
        changes: { type: array, items: { type: string, enum: [code, name, price, cost, packSize, stock] } }
...
  responses:
...
    CommandDone:
      description: |
        Done, now or by an earlier request with the same `operationId`
        (`replayed: true`, and nothing changed this time). `result` is what
        the first run did; `sheet` and `product` are the documents as they
        are now, read after the command.
      content:
        application/json:
          schema:
            type: object
            required: [operationId, replayed, result, product]
            properties:
              operationId: { type: string, format: uuid }
              replayed: { type: boolean }
              result: { $ref: "#/components/schemas/CommandResult" }
              sheet:
                description: Checkout and return only. `null` if the sheet has since been deleted.
                oneOf: [{ $ref: "#/components/schemas/Document" }, { type: "null" }]
              product:
                description: "`null` for an item that isn't in inventory."
                oneOf: [{ $ref: "#/components/schemas/Document" }, { type: "null" }]
...
    CommandRefused:
      description: |
        `bad_request`, nothing changed: a malformed body, a quantity that isn't
        a whole number from 1, money with more than two decimals, a return of
        more than is left out, an item that isn't on the sheet, an item not in
        inventory checked out without `name` and `price`, or an `operationId`
        already used for a different request. Don't retry it unchanged.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    CommandConflict:
      description: |
        `aborted`, nothing changed: the sheet is closed, or the line or item
        kept changing under every retry. The message says which. A retry with
        the same `operationId` is safe.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    BadRequest:
      description: "`bad_request`: malformed JSON, ID, field, query or cursor."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthenticated:
      description: "`unauthenticated`: no token, an expired or ID token. (API Gateway's own 401 has the body `{\"message\":\"Unauthorized\"}`.)"
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: |
        `permission_denied`, with a `reason`: `not_member` (not a member of
        this team, the same answer if the team doesn't exist), `view_only` (a
        viewer tried to write; the web runtime hands this to the app as
        `invalid_argument`, view-only access), `owners_only` (a contributor
        or viewer on an owners-only route) or `team_closed` (a write to a
        team an owner closed, which is read-only). Nothing was read or written.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: "`not_found`: no such document (for a command: no such sheet, or for a stock adjustment no such item)."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooLarge:
      description: "`quota_exceeded`: the document would be over 350,000 bytes (or the body over 1 MB)."
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...